Try to fix casting issues with older versions of GLib.
Try to fix

packet-btatt.c: In function 'btatt_handle_value':
packet-btatt.c:1420: warning: cast to pointer from integer of different size
packet-btatt.c: In function 'btatt_uuid16_value':
packet-btatt.c:1443: warning: cast to pointer from integer of different size

on the 32-bit OS X builder.

Reviewed-on: https://code.wireshark.org/review/9913

+++ epan/dissectors/packet-btatt.c
@@ -1417,7 +1417,7 @@ static gpointer btatt_handle_value(packet_info *pinfo)
     value_data = (guint16 *) p_get_proto_data(pinfo->pool, pinfo, proto_btatt, PROTO_DATA_BTATT_HANDLE);
 
     if (value_data)
-        return GUINT_TO_POINTER(*value_data);
+        return GUINT_TO_POINTER((gulong)*value_data);
 
     return NULL;
 }
@@ -1440,7 +1440,7 @@ static gpointer btatt_uuid16_value(packet_info *pinfo)
     value_data = (guint16 *) p_get_proto_data(pinfo->pool, pinfo, proto_btatt, PROTO_DATA_BTATT_UUID16);
 
     if (value_data)
-        return GUINT_TO_POINTER(*value_data);
+        return GUINT_TO_POINTER((gulong)*value_data);
 
     return NULL;
 }
